requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
python-dotenv
sqlalchemy
psycopg2-binary